+++ cmd/sealer/cmd/apply.go
@@ -15,7 +15,7 @@
 package cmd
 
 import (
-	"github.com/sealerio/sealer/pkg/runtime"
+	"github.com/sealerio/sealer/pkg/runtime/kubernetes"
 	"github.com/spf13/cobra"
 
 	"github.com/sealerio/sealer/apply"
@@ -44,5 +44,5 @@ will apply the diff change of current Clusterfile and the original one.`,
 func init() {
 	rootCmd.AddCommand(applyCmd)
 	applyCmd.Flags().StringVarP(&clusterFile, "Clusterfile", "f", "Clusterfile", "Clusterfile path to apply a Kubernetes cluster")
-	applyCmd.Flags().BoolVar(&runtime.ForceDelete, "force", false, "force to delete the specified cluster if set true")
+	applyCmd.Flags().BoolVar(&kubernetes.ForceDelete, "force", false, "force to delete the specified cluster if set true")
 }
