+++ pkg/runtime/kubernetes/static_files.go
@@ -12,13 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package runtime
+package kubernetes
 
 const (
-	Cluster                = "Cluster"
-	InitConfiguration      = "InitConfiguration"
-	JoinConfiguration      = "JoinConfiguration"
-	ClusterConfiguration   = "ClusterConfiguration"
-	KubeProxyConfiguration = "KubeProxyConfiguration"
-	KubeletConfiguration   = "KubeletConfiguration"
+	AuditPolicyYml = "audit-policy.yml"
 )
+
+// StaticFile :static file should not be template, will never be changed while initialization.
+type StaticFile struct {
+	DestinationDir string
+	Name           string
+}
+
+//MasterStaticFiles Put static files here, can be moved to all master nodes before kubeadm execution
+var MasterStaticFiles = []*StaticFile{
+	{
+		DestinationDir: "/etc/kubernetes",
+		Name:           AuditPolicyYml,
+	},
+}
