+++ pkg/runtime/kubernetes/utils_test.go
@@ -1,0 +1,7 @@
+package kubernetes
+
+import "testing"
+
+func TestVersionCompare(t *testing.T) {
+
+}
