+++ pkg/runtime/utils.go
@@ -19,73 +19,13 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net"
-	"path"
 	"path/filepath"
 	"strings"
 
-	"github.com/sealerio/sealer/utils/exec"
-	"github.com/sirupsen/logrus"
-
 	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/pkg/errors"
 	"github.com/sealerio/sealer/common"
 	osi "github.com/sealerio/sealer/utils/os"
-	"github.com/sealerio/sealer/utils/ssh"
 )
-
-// VersionCompare :if v1 >= v2 return true, else return false
-func VersionCompare(v1, v2 string) bool {
-	v1 = strings.Replace(v1, "v", "", -1)
-	v2 = strings.Replace(v2, "v", "", -1)
-	v1 = strings.Split(v1, "-")[0]
-	v2 = strings.Split(v2, "-")[0]
-	v1List := strings.Split(v1, ".")
-	v2List := strings.Split(v2, ".")
-
-	if len(v1List) != 3 || len(v2List) != 3 {
-		logrus.Errorf("error version format %s %s", v1, v2)
-		return false
-	}
-	if v1List[0] > v2List[0] {
-		return true
-	} else if v1List[0] < v2List[0] {
-		return false
-	}
-	if v1List[1] > v2List[1] {
-		return true
-	} else if v1List[1] < v2List[1] {
-		return false
-	}
-	if v1List[2] > v2List[2] {
-		return true
-	}
-	return true
-}
-
-func GetKubectlAndKubeconfig(ssh ssh.Interface, host net.IP, rootfs string) error {
-	// fetch the cluster kubeconfig, and add /etc/hosts "EIP apiserver.cluster.local" so we can get the current cluster status later
-	err := ssh.Fetch(host, path.Join(common.DefaultKubeConfigDir(), "config"), common.KubeAdminConf)
-	if err != nil {
-		return errors.Wrap(err, "failed to copy kubeconfig")
-	}
-	_, err = exec.RunSimpleCmd(fmt.Sprintf("cat /etc/hosts |grep '%s %s' || echo '%s %s' >> /etc/hosts",
-		host, common.APIServerDomain, host, common.APIServerDomain))
-	if err != nil {
-		return errors.Wrap(err, "failed to add master IP to etc hosts")
-	}
-
-	if !osi.IsFileExist(common.KubectlPath) {
-		err = osi.RecursionCopy(filepath.Join(rootfs, "bin/kubectl"), common.KubectlPath)
-		if err != nil {
-			return err
-		}
-		err = exec.Cmd("chmod", "+x", common.KubectlPath)
-		if err != nil {
-			return errors.Wrap(err, "failed to chmod a+x kubectl")
-		}
-	}
-	return nil
-}
 
 // LoadMetadata :read metadata via ClusterImage name.
 func LoadMetadata(rootfs string) (*Metadata, error) {
